pyproject.toml
[tool.poetry]
name = "remove-myself"
version = "0.1.0"
description = ""
authors = ["Dev Aggarwal <devxpy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.25.0"
google-api-python-client = "^2.96.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
python-decouple = "^3.8"
oauth2client = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
